\name{.compute.gamma.ev.and.pv}
\alias{.compute.gamma.ev.and.pv}
\title{Compute Mixed Gamma Expected Values...}
\usage{.compute.gamma.ev.and.pv(fixed, random, ..., param)}
\description{Compute Mixed Gamma Expected Values
It seems this function is in need of heavy repair.
In particular, the exclusive presence of Inf and
numerically zero results}
\value{a list containing simulated expected values (ev),
predicted values (pv) and warnings accrued from
simulating the predicted values}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{fixed}{fixed values}
\item{random}{random values}
\item{param}{the `parameters' object}}
