\name{describe.logit.mixed}
\alias{describe.logit.mixed}
\title{Describe the `logit...}
\usage{describe.logit.mixed(...)}
\description{Describe the `logit.mixed' model to Zelig}
\value{a list of important information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
