\name{describe.ls.mixed}
\alias{describe.ls.mixed}
\title{Describe a ``ls.mixed'' model to Zelig}
\usage{
  \method{describe}{ls.mixed}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a ``ls.mixed'' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

