\name{describe.poisson.mixed}
\alias{describe.poisson.mixed}
\title{Describe the `poisson...}
\usage{describe.poisson.mixed(...)}
\description{Describe the `poisson.mixed' model to Zelig}
\value{a list of important information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
