\name{model.frame.lmer}
\alias{model.frame.lmer}
\title{Extract the Environment of a `lmer' object...}
\usage{model.frame.lmer(formula, ...)}
\description{Extract the Environment of a `lmer' object}
\value{a `model.frame' object}
\author{Matt Owen and Ferdinand Alimadhi and Delia
\email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a `lmer' object}
\item{...}{ignored parameters}}
