\name{qi.gamma.mixed}
\alias{qi.gamma.mixed}
\title{Compute Quantities of Interest for the Zelig Model gamma...}
\usage{\method{qi}{gamma.mixed}(obj, x, x1=NULL, y=NULL, num=1000, param=NULL)}
\description{Compute Quantities of Interest for the Zelig Model gamma.mixed}
\value{a list of key-value pairs specifying pairing titles of quantities of interest
with their simulations}
\author{Matt Owen and Gregor Gorjanc and Ferdinand Alimadhi and Delia}
\arguments{\item{obj}{a zelig object}
\item{x}{a setx object}
\item{x1}{an optional setx object}
\item{y}{...}
\item{num}{an integer specifying the number of simulations to compute}
\item{param}{a parameters object}}
