\name{qi.logit.mixed}
\alias{qi.logit.mixed}
\title{Compute Quantities of Interest for the Zelig Models logit...}
\usage{qi.logit.mixed(obj, x, x1, y, num=1000, param)}
\description{Compute Quantities of Interest for the Zelig Models logit.mixed and probit.mixed}
\value{a list of key-value pairs specifying pairing titles of quantities of interest
with their simulations}
\arguments{\item{obj}{a zelig object}
\item{x}{a setx object}
\item{x1}{an optional setx object}
\item{y}{ignored until later versions}
\item{num}{an integer specifying the number of simulations to compute}
\item{param}{a parameters object}}
