\name{setup.x.matrix}
\alias{setup.x.matrix}
\title{Setup Matrix...}
\usage{setup.x.matrix(form, x)}
\description{Setup Matrix}
\note{This function is used internally by \code{mixed.zelig}}
\value{a \code{matrix}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{form}{a \code{formula} whose "fixed" terms are used to generate
a matrix for simulating quantities of itnerest}
\item{x}{a \code{setx} object}}
