\name{zelig2probit.mixed}
\alias{zelig2probit.mixed}
\title{Interface between the Zelig Model probit...}
\usage{zelig2probit.mixed(formula, ..., data)}
\description{Interface between the Zelig Model probit.mixed and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{parameter to be passed to the \code{lmer} function}
\item{data}{a data.frame}}
