\docType{package}
\name{ZeligMixed-package}
\alias{ZeligMixed}
\alias{ZeligMixed-package}
\title{Multi-level (Mixed, Hierarchical, etc.)  Zelig Models}
\description{
  \tabular{ll}{ Package: \tab ZeligMixed\cr Version: \tab
  .3\cr Date: \tab 2011-12-12\cr Depends: \tab Zelig,
  lme4\cr License: \tab GPL version 2 or newer\cr URL: \tab
  http://gking.harvard.edu/zelig\cr }
}
\details{
  Add-on pack for Zelig, containing a multi-level models:
  \tabular{ll}{ gamma.mixed: \tab Multi-level Gamma
  Regression \cr logit.mixed: \tab Multi-level Logit
  Regression \cr ls.mixed: \tab Multi-level Least-squares
  Regression \cr poisson.mixed: \tab Multi-level Poisson
  Regression \cr probit.mixed: \tab Multi-level Probit
  Regression \cr }
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai,
  Olivia Lau, and Gary King

  Maintainer: Matt Owen \email{mowen@iq.harvard.edu}
}
\keyword{package}

