\docType{package}
\name{ZeligMultilevel-package}
\alias{ZeligMultilevel}
\alias{ZeligMultilevel-package}
\title{Multi-level Zelig Models}
\description{
  \tabular{ll}{ Package: \tab ZeligMultilevel\cr Version:
  \tab 0.7\cr Date: \tab 2011-12-12\cr Depends: \tab Zelig
  (>= 4.0-11), lme4\cr License: \tab GPL version 2 or
  newer\cr URL: \tab http://gking.harvard.edu/zelig\cr }
}
\details{
  Add-on pack for Zelig, containing multi-level models:
  \tabular{ll}{ gamma.mixed: \tab Multi-level Gamma
  Regression \cr logit.mixed: \tab Multi-level Logit
  Regression \cr ls.mixed: \tab Multi-level Least-squares
  Regression \cr poisson.mixed: \tab Multi-level Poisson
  Regression \cr probit.mixed: \tab Multi-level Probit
  Regression \cr }
}
\author{
  Ferdinand Alimadhi, Delia Bailey Maintainer: Matt Owen
  \email{mowen@iq.harvard.edu}
}
\keyword{Mixed-effect}
\keyword{Multilevel}
\keyword{package}
\keyword{Zelig}

