\name{describe.gamma.mixed}
\alias{describe.gamma.mixed}
\title{Describe the `gamma...}
\usage{describe.gamma.mixed(...)}
\description{Describe the `gamma.mixed' model to Zelig}
\value{a list of important information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
