\name{describe.logit.mixed}
\alias{describe.logit.mixed}
\title{Describe a ``logit.mixed'' model to Zelig}
\usage{
  \method{describe}{logit.mixed}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a ``logit.mixed'' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

