\name{describe.probit.mixed}
\alias{describe.probit.mixed}
\title{Describe the `probit' model to Zelig...}
\usage{describe.probit.mixed(...)}
\description{Describe the `probit' model to Zelig}
\value{a list of important information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
