\name{model.frame.lmer}
\alias{model.frame.lmer}
\title{Extract the Environment of a `lmer' object}
\usage{
  model.frame.lmer(formula, ...)
}
\arguments{
  \item{formula}{a `lmer' object}

  \item{...}{ignored parameters}
}
\value{
  a `model.frame' object
}
\description{
  Extract the Environment of a `lmer' object
}
\author{
  Matt Owen and Ferdinand Alimadhi and Delia
  \email{mowen@iq.harvard.edu}
}

