\name{setup.x.matrix}
\alias{setup.x.matrix}
\title{Setup Matrix}
\usage{
  setup.x.matrix(form, x)
}
\arguments{
  \item{form}{a \code{formula} whose "fixed" terms are used
  to generate a matrix for simulating quantities of
  itnerest}

  \item{x}{a \code{setx} object}
}
\value{
  a \code{matrix}
}
\description{
  Setup Matrix
}
\note{
  This function is used internally by \code{mixed.zelig}
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

