\name{zelig2poisson.mixed}
\alias{zelig2poisson.mixed}
\title{Interface between the Zelig Model poisson...}
\usage{zelig2poisson.mixed(formula, family=gaussian, ..., data)}
\description{Interface between the Zelig Model poisson.mixed and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{}
\item{data}{a data.frame}}
