\name{zelig2probit.mixed}
\alias{zelig2probit.mixed}
\title{Interface between the Zelig Model probit.mixed and
the Pre-existing Model-fitting Method}
\usage{
  zelig2probit.mixed(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{parameter to be passed to the \code{lmer}
  function}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model probit.mixed and the
  Pre-existing Model-fitting Method
}

